### 2023-10-15 #613 @datapumpernickel
comtradr: API wrapper for UN Comtrade data
Submitting Author Name: Paul Bochtler
Submitting Author Github Handle: @datapumpernickel
Repository:  https://github.com/ropensci/comtradr
Version submitted: 0.4.0 (not yet released)
Submission type: Standard

Version accepted: TBD

---

```
Package: comtradr
Title: Interface with the United Nations Comtrade API
Version: 0.4.0.0
            person("Paul", "Bochtler",
              email = "paulbochtler.gh@gmail.com",
              role = c("aut", "cre","cph"),
            person("Harriet", "Goers",
              email = "hgoers@umd.edu",
              role = c("aut")),
            person("Chris", "Muir",
              email = "chrismuirRVA@gmail.com",
              role = c("aut")),
            person("Alicia", "Schep",
                  role = "rev",
                  comment = c(ORCID = "0000-0002-3915-0618",
                              "Alicia reviewed the package for rOpenSci,
                              see https://github.com/ropensci/onboarding/issues/141")),
            person("Rafael", "Hellwig",
                  role = "rev",
                  comment = c(ORCID = "0000-0002-3092-3493",
                              "Rafael reviewed the package for rOpenSci,
                              see https://github.com/ropensci/onboarding/issues/141")),
            person("Juergen", "Amann", role=c("ctb")))
Description: Interface with and extract data from the United Nations Comtrade
  API <https://comtradeplus.un.org/>. Comtrade provides country level shipping
  data for a variety of commodities, these functions allow for easy API query
  and data returned as a tidy data frame.
    lifecycle,
    fs,
    askpass,
    cli,
    httr2,
    poorman,
    purrr
NeedsCompilation: no
Maintainer: Paul Bochtler <paulbochtler.gh@gmail.com>
    covr,
    httptest2,
    vcr

```

The package leverages httr2 to extract trade data from the UN Comtrade API, which is one of the standard sources for trade data world wide. See [here](https://comtradeplus.un.org/) for more info on the API.

The package can be used for econometric research into the trade relations of countries, dependencies between countries and many other use cases for trade data. The target audience are students and scientists, as well as practitioneers in the field of econometrics, political science among others.

As far as I am aware, there is no other package that wraps the UN Comtrade API.

The package has previouslz been reviewed by rOpenSci and is already part of the rOpenSci suite. However, since the API underwent fundamental changes, so did most of the functions of the package. Hence, we asked, whether we could get another review. This inquiry was answered positively here: https://github.com/ropensci/software-review-meta/issues/100

As far as I am aware, the package currently passes all the pkgcheck checks.
