### 2023-11-06 #613 @ernestguevarra
Hi @datapumpernickel. Just a quick note to say that I am almost done with a comprehensive pass at reviewing your package! Sorry that I have been delayed. When I accepted the review role, I forgot where I should look to see the review process and my GitHub notifications were getting archived on my email.

Otherwise, I will be posting my review notes here in the next couple of hours.
