### 2023-11-06 #613 @ernestguevarra
@datapumpernickel, here is my review:

---

#### Test installation

Both local install and remote install (via GitHub) of the package proceeded without issues and as expected/documented.

#### Check package integrity

* No issues on `{devtools}` checks. Check results were:

```R
── R CMD check results ──────────────────────────────────────────────────────────────────────────── comtradr 0.4.0.0 ────
Duration: 19.1s

0 errors ✔ | 0 warnings ✔ | 0 notes ✔
```

* No issues on `{devtools}` test. Test results were:

```R
══ Results ══════════════════════════════════════════════════════════════════════════════════════════════════════════════

```

* Following issues raised on `{goodpractice}` test/check:

```R
── GP comtradr ──────────────────────────────────────────────────────────────────────────────────────────────────────────

  ✖ write unit tests for all functions, and all package code in general. 79% of code lines are
    covered by test cases.

    R/ct_build_request.R:52:NA
    R/ct_check_params.R:296:NA
    R/ct_check_params.R:300:NA
    R/ct_check_params.R:301:NA
    R/ct_check_params.R:302:NA

  ✖ avoid long code lines, it is bad for readability. Also, many people prefer editor windows that
    are about 80 characters wide. Try make your lines shorter than 80 characters

    data-raw/cmd_codes.R:7:81
    data-raw/country_codes.R:7:81
    data-raw/country_codes.R:22:81
    data-raw/country_codes.R:53:81
    data-raw/DATASET.R:11:81

  ✖ fix this R CMD check NOTE: Note: found 11 marked UTF-8 strings
─────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
```

- `r emo::ji("check")` On unit tests for all functions, as per [rOpenSci Packaging Guide](https://devguide.ropensci.org/building.html) documentation on [tests](https://devguide.ropensci.org/building.html#testing), a minimum coverage of 75% is required so this should be OK. I looked at a few of the lines that are not covered and it seems to be those that tend to be hard to test with CRAN and rhub in mind. Maybe for the sake of being thorough and complete, a final look at non-covered lines of code can be looked at and see if some of them can be covered with just some adjustments to existing test suites and/or adding straightforward tests to get them covered.

- `r emo::ji("warning")` On avoiding long lines, a set of these are from your scripts inside the `data-raw` directory which processes your reference data that you use in the package. I am not 100% sure about this but for those, I think the strict linting may not be necessary. I think the `data-raw` directory can be skipped in the lint checks. For others, the lines that exceed tend to be long URLs and long messages that guide the user regarding a warning or an error. It is always challenging for these as URLs can't really be split/wrapped into more than one line whilst error/warning/messages may not show up on console in a tidy/pretty way. For this, I might suggest adding `# nolint` at the end of the long line to quiet down these messages from the good practice check.

- `r emo::ji("question")` On finding 11 marked UTF-8 strings, this NOTE doesn't come up on `devtools::check()` so maybe it comes out on rhub checks? I think these are likely coming from the datasets included in the package particulary country names with accented characters. I am not fully certain what best practice is for this. I think it is ideal that the actual names provided by Comtrade are used in the dataset but to be able quiet down this issue, it might be needed to convert text to ASCII. For this `stringi::stri_enc_toascii()` might be useful.

#### Check package metadata files

* on `devtools::spell_check()`, I see:

```R
  WORD          FOUND IN
️           README.md:49,102,124
              README.Rmd:40,83,106
accomodate    README.md:28
              README.Rmd:30
arg           ct_commodity_lookup.Rd:40
              NEWS.md:26,56,57,58
              comtradr.Rmd:98,194,211
args          ct_commodity_lookup.Rd:45
              NEWS.md:50
ASEAN         country_codes.Rd:16
CMD           README.md:10
              README.Rmd:18
Codecov       README.md:12
              README.Rmd:19
comtrade      ct_build_request.Rd:10
              ct_perform_request.Rd:10,17
              get_primary_comtrade_key.Rd:5,10
              set_primary_comtrade_key.Rd:10,13
              README.md:91
              README.Rmd:75
              comtradr.Rmd:61
Comtrade      comtradr-package.Rd:7,11
              country_codes.Rd:17,18,28
              ct_commodity_lookup.Rd:5,52,54,57
              ct_get_data.Rd:5,53,69,77
              ct_perform_request.Rd:5,20
              get_primary_comtrade_key.Rd:13
              set_primary_comtrade_key.Rd:5,16
              title:1
              description:1,2
              NEWS.md:29,54,77
              README.md:15,16,24
              README.Rmd:22,22,28
              comtradr.Rmd:29,29,119,183,203,211
COMTRADE      get_primary_comtrade_key.Rd:10,13
              set_primary_comtrade_key.Rd:16
env           get_primary_comtrade_key.Rd:13
              set_primary_comtrade_key.Rd:16
func          NEWS.md:26,32,34,36
github        comtradr-package.Rd:33,34
https         comtradr-package.Rd:33,34
httr          ct_build_request.Rd:17,20
              ct_get_data.Rd:47,66
              ct_process_response.Rd:10,20
ℹ️        README.md:95
              README.Rmd:77
              comtradr.Rmd:63
iso           ct_country_lookup.Rd:16
              ct_process_response.Rd:20
json          ct_perform_request.Rd:17
              ct_process_response.Rd:20
              NEWS.md:34
natively      comtradr.Rmd:192
onboarding    comtradr-package.Rd:33,34
ORCID         comtradr-package.Rd:23,33,34
Readme        NEWS.md:4
repo          README.md:51
              README.Rmd:42
ropensci      comtradr-package.Rd:33,34
              README.md:150
              README.Rmd:130
rOpenSci      comtradr-package.Rd:33,34
              README.md:9
              README.Rmd:17
stringified   NEWS.md:36
testthat      NEWS.md:83,99
twi           ct_pretty_cols.Rd:9
useability    README.md:36
              README.Rmd:32
```

* For the README, please address some of the non-false positive results produced by `devtools::spell_check()`. After editing the spelling of non-false positive results, please run `usethis::use_spell_check()` which will allow you to record the words that have been false-positvely flagged as misspelled into a `WORDLIST` kept in the `inst` folder. Once they are on this list, they will not anymore be flagged as misspelled. Please note that when running `usethis::use_spell_check()`, language will default to `en-US` and will be added to your DESCRIPTION file. Also, the `spelling` package will be added in the `Suggests`.

* For the README, in the sentence of the first paragraph, it might be good to make this consistent with the text used in the DESCRIPTION file so something like "R package for interfacing with and extracting data from ..." to make the two consistent and coherent with each other. Otherwise, README clearly states the problems the software is designed to solve and its target audience

* Maybe organise the entries on the DESCRIPTION file more coherently. For example, move the line for Maintainer right after the line/lines for Authors@R and the entries for Suggests be put right after the entries for Imports. The DESCRIPTION file will still work appropriately without changing these but the file will read better if this change is made.

* No issues with the NAMESPACE

#### Check documentation

* Key functions of the package have help files that are generally appropriate and describe in fair detail what the function does and what it needs to produce such output.

* Supporting functions or utility functions, however, tend to be less specific in the description of their functionality. In most cases, these are functions that users will most likely not use directly as they are used by the package within the key/main functions. As such, it might be good to note this with the description of the function.

* Some other functions are legacy functions that are lined up for deprecation or have been superseded. For these, it might be good to have text in the description that indicates what the deprecation plan is for these functions.

* In the package website built using `{pkgdown}`, the references page for the function help files might benefit from some organisation of the functions. Groupings will depend on how you want to organise them but generally, they are groupings based on related functionalities. Also, this a good way to group data elements, legacy functions and functions meant for deprecation. This can be done by just creating `_pkgdown.yml` file by issuing the following command `pkgdown::use_pkgdown()`. Once that file is created, you can edit/customise it in many ways including grouping of the reference pages for each function. To see how this is done, look at the `_pkgdown.yml` file of the `{pkgdown}` package website - https://github.com/r-lib/pkgdown/blob/main/pkgdown/_pkgdown.yml.

* The package currently has one vignette which gets produced as the **Getting Started** vignette. It expands the content of the README with more focused examples. This is good as it re-emphasises the main use case of `{comatradr}` package.

* It might be good to consider adding at least one other vignette aimed at those who have used or are familiar with the earlier iteration of the package and explain/differentiate through a specific use case (i.e., retrieving data) the previous approach as compared to the current approach. In this, links to the new data structure for Comtrade provided by the UN will be helpful as this will give the user an idea of what new fields they can expect in the output of the function that retrieves data.

* Another vignette that might be useful specifically for those who used the previous version to bulk download data from Comtrade is about what users need to consider if they are to going to do a bulk download and what features of the new package help with this (limits to number of years, limits to number of characters in API calls, throttling of the API calls). These elements are in different places in the documentation but might be good to be put together in a vignette using a worked example of a use case of performing a bulk download (i.e., multiple years, lots of commodities).

#### Test functionality

I have been using the new `{comtradr}` package since June of this year for a project I am involved in that is downloaded data on 963 commodities for all countries from 2000 to 2023. The team I am working with had original/previous code that used the `{tradestatistics}` package to do this at a time of the previous API specification for the UN Comtrade. When I joined the team for the second phase of the project, the API changed and the both the `{tradestatistics}` package and the legacy `{comtradr}` package underwent changes to accommodate the new API specifiations. For the purposes that we need it for (bulk download), the updated `{comtradr}` package was fit for our needs as its main function for getting data factored in all the rate-limits that the new API has put in place and also includes API call throttling (which I really think is very well thought of and well implemented). Whilst there are still sometimes 500 error issues due to the UN Comtrade server timing out (probably due to have query loads), the `{comtradr}` package has really been able perform its single most important functionality of retrieving data from UN Comtrade.

So, based on this experience since June, I can say that I've been able to push the `{comtradr}` functionality to its limits and given restrictions on a free API account, we are happy with what we are able to retrieve using this package.

A few things that maybe considered by the author/maintainer (probably as a new feature in upcoming updates) are:

* Better mechanism to catch the infrequent 500 errors that popup secondary to the UN Comtrade server timing out and an appropriate error message that gives information regarding this. Because of how good `{comtradr}` deals with the throttling, if you are doing a reasonbly sized bulk download, the 500 errors don't come up and the waiting time messaging on console is informative enough. But for instances as I described above for what we used it for, it might be useful to have a bit more information messaging when indeed a 500 error comes up.

* Maybe consider adding an argument that allows user to specify how much to throttle the API calls. From what I can see in the functions of the package, the throttling is fixed to 6 calls per minute. In our use case above, we would have liked to have had the capability to throttle the calls a little bit more.

* Maybe consider functions that deal with the added parameters for paid accounts. I think this is not easy to implement as it would mean having a paid account yourself so this might be down the priority list. But I think it would be good in the future to have these particularly for the bulk download capabilities of a paid for account.
