### 2023-11-06 #613 @ernestguevarra
@datapumpernickel you may notice that a big chunk of my review has been on the documentation. As I was doing the review, I forked your package and have been making edits on the docs/help files/vignettes as I was reviewing them. I am happy to make a pull request with my edits so you can have a look at what these changes may look like. Let me know if you think this would be helpful.

Otherwise, thank you for your good work on getting `{comtradr}` revived. It's great work that you have put in and I think we can polish it even more! Well done!
