### 2023-11-07 #613 @datapumpernickel
Hi @ernestguevarra!

first of all, thank you again for your thorough review and helpful comments - this is great! I am also really happy to hear that you were able to already put the package to use and found it useful.

1) Yes, I would be very happy to have a pull request with your edits and corrections and merge it into the package. Maybe you could pull into the `dev` branch. Thanks!

2) I will subsequently work through your comments and either make the respective changes or add it to the list of future features! Already now I can say, allowing the user to control throttling is a great idea and easy to implement. I was hesitant to add more arguments to the already long list, but I totally see how that is necessary. Paid features are indeed on the list for future development, I hope to have premium access next year.
