### 2023-11-07 #613 @ernestguevarra
@datapumpernickel I will finalise my edits today on the package (mainly documentation) and will make a PR to `dev` as recommended.

Also, you can just use the PR as a reference for what I am suggesting to change and you can implement the changes yourself as you update the package. I think sometimes that might be easier so you have full control of the changes. So, do with it as you deem best and most efficient for your workflow.
