### 2023-11-10 ropensci/comtradr#69 @datapumpernickel
Review - planned changes
- [x] add `# nolint` to long lines
- [x] change country names to ASCII (slight preference to keep original names)
  - have decided to keep it in line with original names from comtradr, hoping it wont cause more issues with any checks.
- [x] correct spell check and add 'wrong' words to list of non-checked words
- [x] make README consistent with DECRIPTION file
- [x] reorganise entries in DESCRIPTION (e.g. suggests after Imports and Maintiner after line for authors
- [x] add some more help info on supporting functions that are not used by user or specify why there is less help docu there
- [x] reorganise functions in pkgdown site by grouping them by functionality...
- [x] better explanation of `comtradr::country_codes$iso_3` and `ct_get_ref_table()` which is helpful for getting the codes right
- [x] one more vignette for transitioning from previous package to new package
- [x] specifically one vignette for querying larger amounts of data and bulk download (or workarounds for this functionality)
- [x] add more explicit description of how to get the API key
- [x] review documentation to respect character limit of 80 per line...

- [x] allow user to control throttling
- [x] change rate limit to 100k not 250k
- [x] provide mode_of_transport as text, not as ID, as has been implemented with all other arguments as well, that seems to be more straight forward.
- [x] catch 500 error messages and maybe increase throttling in this case
  - could not replicate, maybe in an earlier iteration of the package this happened, but now the package should automaticlly back-off even longer than specified, if the Comtrade API blocks a request, so there really should not be 500 errors.
- [x] confusion around leaving argument blank, specifying NULL and the default as per the current documentation. Maybe change default to align with `NULL`? This however, results in a lot of values being returned by default, when most users want the "total", e.g. for customs.
- [x] provide `all` for rest of arguments that do not allow this yet
