### 2023-11-10 03141fd
allow user to control throttling manually
