+++ .github/workflows/test-coverage.yaml
@@ -2,9 +2,9 @@
 # Need help debugging build failures? Start at https://github.com/r-lib/actions#where-to-find-help
 on:
   push:
-    branches: [main, master,api-update]
+    branches: [main, master,dev]
   pull_request:
-    branches: [main, master,api-update]
+    branches: [main, master,dev]
 
 name: test-coverage
 
