### 2023-11-11 5eb0418
made readme consistent with DESCRIPTION
