### 2023-11-11 47da2c0
documented internal functions a bit more clearly, notably adding which functions call them
