### 2023-11-12 bf01c46
added examples to internal functions and added comtradr::: so the examples woudl work
