+++ R/ct_perform_request.R
@@ -15,7 +15,7 @@
 #'             end_date = '2019',
 #'             flow_direction = 'import')
 #' req <- httr2::last_request()
-#' resp <- ct_perform_request(req, requests_per_second = 10/60, verbose = FALSE)
+#' resp <- comtradr:::ct_perform_request(req, requests_per_second = 10/60, verbose = FALSE)
 #' @inheritParams ct_get_data
 ct_perform_request <- function(req, requests_per_second, verbose = FALSE) {
 
