+++ man/ct_perform_request.Rd
@@ -30,6 +30,6 @@ ct_get_data(commodity_code = NULL,
             end_date = '2019',
             flow_direction = 'import')
 req <- httr2::last_request()
-resp <- ct_perform_request(req, requests_per_second = 10/60, verbose = FALSE)
+resp <- comtradr:::ct_perform_request(req, requests_per_second = 10/60, verbose = FALSE)
 \dontshow{\}) # examplesIf}
 }
