+++ data-raw/DATASET.R
@@ -8,7 +8,7 @@ library(usethis)
 # getting comtrade data ---------------------------------------------------
 
 ## getting list of reference tables
-response <- httr2::request('https://comtradeapi.un.org/files/v1/app/reference/ListofReferences.json') |>
+response <- httr2::request('https://comtradeapi.un.org/files/v1/app/reference/ListofReferences.json') |> # nolint
   httr2::req_perform()
 
 ## getting data from response of list of tables
@@ -17,7 +17,8 @@ list_of_datasets <- response |>
   purrr::pluck(1)
 
 ## getting date of last modification from list of tables
-last_modified <- httr2::resp_header(header = "Last-Modified", resp = response) |>
+last_modified <- httr2::resp_header(header = "Last-Modified",
+                                    resp = response) |>
   stringr::str_extract(pattern = '(\\d{2} [a-zA-Z]+ \\d{4})') |>
   as.Date(format = "%d %b %Y")
 
@@ -25,15 +26,16 @@ last_modified <- httr2::resp_header(header = "Last-Modified", resp = response) |
 list_of_datasets$last_modified <- last_modified
 
 ## changing colon to underscore in dataset names
-list_of_datasets$category <- stringr::str_replace_all(list_of_datasets$category,':',"_") |>
+list_of_datasets$category <- stringr::str_replace_all(list_of_datasets$category,
+                                                      ':',"_") |>
   tolower()
 
 ## save list of datasets
 save(list_of_datasets, file = 'inst/extdata/list_of_datasets.rda')
 
-i <- 1
 ## loop over all datasets (for loop, because it is readable,
-## no need for speeding this up with more complicated lapply or more dependencies)
+## no need for speeding this up with more complicated lapply
+## or more dependencies)
 for(i in seq_along(list_of_datasets$category)){
   ## define the valid commodity codes that we need
   valid_cmd_datasets <- c('cmd_hs', 'cmd_s1', 'cmd_s2', 'cmd_s3', 'cmd_s4',
@@ -51,7 +53,8 @@ for(i in seq_along(list_of_datasets$category)){
     data <- response |>
       httr2::resp_body_json(simplifyVector = T)
 
-    last_modified <- httr2::resp_header(header = "Last-Modified", resp = response) |>
+    last_modified <- httr2::resp_header(header = "Last-Modified",
+                                        resp = response) |>
       stringr::str_extract(pattern = '(\\d{2} [a-zA-Z]+ \\d{4})') |>
       as.Date(format = "%d %b %Y")
 
@@ -60,7 +63,8 @@ for(i in seq_along(list_of_datasets$category)){
     result$last_modified <- last_modified
 
     readr::write_rds(result, "xz",
-                     file = paste0('inst/extdata/',list_of_datasets$category[i],'.rds'))
+                     file = paste0('inst/extdata/',
+                                   list_of_datasets$category[i],'.rds'))
   } else if(list_of_datasets$category[i] %in% valid_country_datasets) {
     response <- httr2::request(list_of_datasets$fileuri[i]) |>
       httr2::req_perform()
@@ -68,7 +72,8 @@ for(i in seq_along(list_of_datasets$category)){
     data <- response |>
       httr2::resp_body_json(simplifyVector = T)
 
-    last_modified <- httr2::resp_header(header = "Last-Modified", resp = response) |>
+    last_modified <- httr2::resp_header(header = "Last-Modified",
+                                        resp = response) |>
       stringr::str_extract(pattern = '(\\d{2} [a-zA-Z]+ \\d{4})') |>
       as.Date(format = "%d %b %Y")
 
@@ -99,7 +104,8 @@ for(i in seq_along(list_of_datasets$category)){
     result$last_modified <- last_modified
 
[…]
