### 2023-11-13 6eb6616
corrected spell checking, added to no spell check non-positive finds
