+++ vignettes/comtradr.Rmd
@@ -203,7 +203,7 @@ In addition to these rate limits, the API imposes some limits on parameter combi
 As explained previously, making API calls with `comtradr` often requires the user to query the commodity reference table (this is done using functions `ct_commodity_lookup`). These reference tables are generated by the UN Comtrade, and are updated roughly once a year. Since they're updated infrequently, the tables are saved as cached data objects within the `comtradr` package, and are referenced by the package functions when needed. 
 
 The function features an `update` argument, that checks for updates, downloads the new tables if necessary and makes them available during the current R session. It will also print a message indicating whether updates were found, like so:
-```{r}
+```{r, eval = F}
 ct_commodity_lookup('tomato',update = T)
 ```
 If any updates are found, the message will state which reference table(s) were updated.
