### 2023-11-17 ed69e72
include headers in mock dir for goods
