### 2023-11-17 0a9c660
updated service endpoint mock_dir with valid data
