### 2023-11-17 2ba899a
fix error, that B4 classification was wrongly coded under services instead of goods
