+++ tests/testthat/test-ct_check_parameters.R
@@ -17,7 +17,7 @@ test_that("check_freq returns correct frequency codes and handles invalid inputs
 # Test 'check_clCode' function
 test_that("check_clCode returns correct classification codes and handles invalid inputs", {
   expect_equal(check_clCode("C", "HS"), "HS")
-  expect_equal(check_clCode("S", "B4"), "B4")
+  expect_equal(check_clCode("C", "B4"), "B4")
   expect_error(check_clCode("C", "ISIC"), "`commodity_classification` must be one of")
   expect_error(check_clCode("S", "HS"), "`commodity_classification` must be one of")
 })
