### 2023-12-09 #613 @potterzot
- **Briefly describe any working relationship you have (had) with the package authors.** No working relationship.

---

First, my deep apologies on the delay of my review. Second, thank you for putting together such a clear and carefully written package. It was a pleasure to review. My comments are mostly limited to a set of issues that I ran into in the documentation and functionality of the `World` and `All` and `all` values for some `ct_get_data()` parameters.

I have included modifications to the documentation in a PR that I will make a request for once this is submitted, but these are just suggestions of course.

No issues with installation.

#### Package Checks

`devtools::check()` and `devtools::test()` present no issues.

- It took me a little while to figure out where to get the proper API. First that you can't just select the API that you want and instead have to type it in, and then whether I need a Free API or a Premium API and which API I need to subscribe to. I recommend adding a bit more to suggest using the "comtradr-v1" free API at  [README.Rmd:57](https://github.com/ropensci/comtradr/blob/c44b45ce96b119ddff9c267da16ec994525e2c01/README.Rmd#L57) and [comtradr.Rmd:45](https://github.com/ropensci/comtradr/blob/c44b45ce96b119ddff9c267da16ec994525e2c01/vignettes/comtradr.Rmd#L45). I know that this risks being out of date any time COMTRADE updates its API site, but those updates may necessitate package updates anyway so maybe this is okay?
- Examples in the README and Vignette work with no issues.
- I wanted to download all imports from all partners for the USA based on the vignette example. Removing the specific countries in the `partner` parameter in `ct_get_data()` defaults to "World" as noted in the documentation, but it looks like "all" is an option as well that will successfully get all partners. It would be worth updating the documentation to reflect "all" as an option since it isn't in the `comtradr::country_codes$iso_3`. The `reporter` parameter might be clarified similarly. In my PR I've modified the documentation for these parameters with a suggestion of "Partner ISO3 code(s), `all`, or `NULL`, ..." .
- Related to that, the vignette mentions that "all" is a convenience implemented by `comtradr`, which is great. Only the documentation in the vignette uses "All" instead of "all", which is what the function `ct_check_params()` requires [here](https://github.com/ropensci/comtradr/blob/c44b45ce96b119ddff9c267da16ec994525e2c01/R/ct_check_params.R#L400).
- Also, [comtradr.Rmd:192](https://github.com/potterzot/comtradr/blob/c44b45ce96b119ddff9c267da16ec994525e2c01/vignettes/comtradr.Rmd#L192) refers to `All` but it is implemented in the lower case and for all of `reporter`, `partner`, and `partner2` parameters.
- And as well, [comtradr.Rmd:]() says that leaving `partner` as `NULL` will result in "all" being used but currently it results in "World" being used. This 'total' default is true for the `mode_of_transport` and `customs_code` parameters as well, but the vignette suggests it should return all values, not the total.
- I suggest adding a reference to comtradr::country_codes$iso_3 in [comtradr.Rmd:117](https://github.com/ropensci/comtradr/blob/c44b45ce96b119ddff9c267da16ec994525e2c01/vignettes/comtradr.Rmd#L117C138-L117C138).
- in `ct_get_data()`, `mode_of_transport` documentation does not make it clear that the `id` is needed and not the `text. For example, for `Air` it must be `ct_get_data(mode_of_transport = 1000, ...)`, not `ct_get_data(mode_of_transport = Air, ...)`. Making the parameter documentation "Mode of Transport ID code..." would clear that up.
- Much of the function documentation exceeds the 80 character line limit. I know this is a pain because I had to change all of my documentation to fit this limit when my own package was reviewed! But it is a style choice of ropensci.

- Unlike `partner`, the `mode_of_transport` and `customs_code` parameters in `ct_get_data()` do not take `all` as a valid parameter value. Perhaps this could be added?

- No issues with tests. These are nicely written.
