### 2023-12-12 #613 @datapumpernickel
Hi @noamross @ernestguevarra  and @potterzot,

thanks to all again for your valuable time and the thorough review! Your comments, additions and tests have already been of great help!

I have made an issue with all the requested changes and am tracking which ones I have already implemented there: https://github.com/ropensci/comtradr/issues/69

Some of the larger documentation tasks (new vignettes for transitioning from older iteration and on how to loop for bulk data) might take me a little longer, but some I might get done in the next week. I will report back once I am done with all the changes.

There is two issues, where I think I could need some more input:

#### Question 1
Currently, I use the `...` to allow for users to add arbitrary parameters to the URL call in order to have a certain flexibility, should the API add new parameters that I have not implemented yet. These are passed on, as is without any checking. Using `...` has been very easy to implement, but I had at least one user, who had the following issue:
Upon specifying a call, like:

``` r
comtradr::ct_get_data(reporter = "DEU",
                      patner ='ARG',
                      start_date = 2020,
                      end_date = 2020)
```

The API did not work as expected. Instead of returning partner Argentina, it returned partner "World" without an error, because the mistyped `patner` argument was just catched by the `...`. So this seems to be very error prone.
I am now tending towards just excluding the `...`, because the main functionality is covered and should there be new parameters that are really urgent, I am fairly confident we can implement them relatively quickly.

However, do you have any other idea on how to allow the passing on of arbitrary parameters to the function? e.g. an argument that takes a named list? I am a bit our of my depth here.

#### Question 2 - Confusion around NULL, "all" and default values.

As has been noted by @potterzot there is some confusion around this. Here is whats happening:

##### Argument in Blank
``` r
comtradr::ct_get_data(reporter = "DEU",
                      start_date = 2020,
                      end_date = 2020)
```
--> This returns Germanys trade with "World", as the argument is unspecified, it defaults to "World".

##### Argument set to NULL
``` r
comtradr::ct_get_data(reporter = "DEU",
                      partner = NULL,
                      start_date = 2020,
                      end_date = 2020)
```
--> This returns Germanys trade with all possible partners, _notably, including World!_

##### Argument set to 'all'
``` r
comtradr::ct_get_data(reporter = "DEU",
                      partner ='all',
                      start_date = 2020,
                      end_date = 2020)
```
This returns Germanys trade with all partner countries that are not groups (e.g. ASEAN is not included, World is not included). This is what you would use, if you would want to aggregate the World values yourself.

When I implemented this, it seemed clear to me, but I can now see how this confuses people. I am thinking about changing the default from "World" to NULL, because it is more intuitive in the documentation to think about leaving it in blank as similar to as setting it to NULL. However, for the parameters custom_code, and mode_of_transport, it is quite confusing, when all of a sudden for one trade relationship you get a bunch of weird entries with different modes of transport and custom codes, when in 99% of all cases, what you want is only the total (Data is realtively new in comtrade and not that reliable yet either). Hence the default "total".

Do you have any ideas/suggestions on how to best solve this?

Maybe I should just include it as clear as here in the documentation somewhere?

Thank you again for your time, really highly appreciated!
