### 2023-12-12 #613 @noamross
@datapumpernickel This a new one for us, but I would say yes, keep the previous reviewers, just as you retain the original author.  Their contribution to the lineage of the package remains, even if much of the code they reviewed does not.
