### 2023-12-12 #613 @noamross
Q1: I think an `extra_params` argument that takes a list of additional parameters would serve.  I think having this option is useful from the perspective of reducing future maintainer burden.

Q2: I agree that NULL is a common default value. If NULL is an option, I would have it be the default, and not an uncommonly used option.  I would suggest (a) the default return smaller data, and (b) that all the options be equivalent. So in this case my suggestion is to retain 'world' as default and remove NULL as an option.  Instead come up with another descriptive named option like `'everything'` that does the equivalent of NULL.
