### 2023-12-15 3458392
edited the help file in utils in order to specify all possible values when querying tables from ct_get_ref_table
