+++ R/utils.R
@@ -38,9 +38,38 @@ get_primary_comtrade_key <- function() {
 
 #' Get reference table from package data
 #'
-#' The first time, the function will read from disk, the second time from the environment. In the case of a necessary update the new data will be saved to the environment for the current session.
-#' You can use this table to look at the reference tables and if necessary extract respective classification codes by hand. In general we would recommend the function `ct_commodity_lookup` for this purpose. It uses the present function in the backend.
-#' @param dataset_id The dataset ID, which is either partner, reporter or a valid classification scheme.
+#' The first time, the function will read from disk, the second time from the
+#' environment. In the case of a necessary update the new data will be saved
+#' to the environment for the current session.
+#' You can use this table to look at the reference tables and if necessary
+#' extract respective classification codes by hand. In general we would
+#' recommend the function `ct_commodity_lookup` for this purpose.
+#' It uses the present function in the backend.
+#'
+#' @details The function allows you to query most possible input parameters
+#' that are listed by the Comtrade API. The following dataset_ids are permitted:
+#' * Datasets that contain codes for the `commodity_code` argument. The name is
+#' the same as you would provide under `commodity_classification`.
+#'   * 'HS' This is probably the most common classification for goods.
+#'   * 'B4'
+#'   * 'B5'
+#'   * 'EB02'
+#'   * 'EB10'
+#'   * 'EB10S'
+#'   * 'EB'
+#'   * 'S1'
+#'   * 'S2'
+#'   * 'S3'
+#'   * 'S4'
+#'   * 'SS'
+#'
+#' * 'reporter'
+#' * 'partner'
+#' * 'mode_of_transport'
+#' * 'customs_code'
+#'
+#' @param dataset_id The dataset ID, which is either partner,
+#' reporter or a valid classification scheme.
 #' @inheritParams ct_get_data
 #' @export
 #' @returns a tidy dataset with a reference table
@@ -70,9 +99,8 @@ ct_get_ref_table <- function(dataset_id, update = FALSE, verbose = FALSE) {
     'SS'    = 'cmd_ss'   ,
     'reporter'    = 'reporter'   ,
     'partner'    = 'partner',
-    'mot'    = 'mot',
-    'customs'    = 'customs'
-
+    'mode_of_transport'    = 'mot',
+    'customs_code'    = 'customs'
   )
 
   ## check dataset id for valid values
