### 2023-12-15 900bc92
renamed some arguments to ct_get_ref_table so they are more intuitive
