+++ R/zzz.R
@@ -30,8 +30,8 @@ assign('S2', NULL, envir = ct_env)
 assign('S3', NULL, envir = ct_env)
 assign('S4', NULL, envir = ct_env)
 assign('SS', NULL, envir = ct_env)
-assign('mot', NULL, envir = ct_env)
-assign('customs', NULL, envir = ct_env)
+assign('mode_of_transport', NULL, envir = ct_env)
+assign('customs_code', NULL, envir = ct_env)
 assign('reporter', NULL, envir = ct_env)
 assign('partner', NULL, envir = ct_env)
 assign('list_of_datasets', NULL, envir = ct_env)
