### 2023-12-15 234d616
new help file for changed arguments to ct_get_ref_table
