### 2023-12-15 5efeca4
added a paragraph on using ct_get_ref_table in the vignette
