### 2023-12-15 f1cfaa4
implementing mode of transport as text, instead of id, extra_params in check_params, deleteed unnecessary parameteres and arguments from ct_check_parameter
