### 2023-12-15 006abbc
reflect changes in arguments in help file of ct_get_data, ad more desctiption for ct_get_ref_table(), added linebreaks in documentation
