+++ R/ct_get_data.R
@@ -1,39 +1,71 @@
 #' Get trade data from the UN Comtrade API
 #'
-#' This function queries the UN Comtrade API to retrieve international trade data.
-#' It allows for detailed specification of the query, including the type of data (goods or services),
-#' frequency (annual or monthly), commodity classification, flow direction, and more.
-#' By providing `NULL` for certain parameters, you can query all possible values.
-#' The function is opinionated in that it already verifies certain parameters for you and
-#' is more than a pure wrapper around the API.
+#' This function queries the UN Comtrade API to retrieve
+#' international trade data.
+#' It allows for detailed specification of the query,
+#' including the type of data (goods or services),
+#' frequency (annual or monthly), commodity classification,
+#' flow direction, and more.
+#' By providing `NULL` for certain parameters,
+#' you can query all possible values.
+#' The function is opinionated in that it already verifies certain parameters
+#' for you and is more than a pure wrapper around the API.
 #'
 #' @details
-#' The UN Comtrade database provides a repository of official international trade statistics
-#' and relevant analytical tables. It contains annual trade statistics starting from 1988
+#' The UN Comtrade database provides a repository of official international
+#' trade statistics and relevant analytical tables.
+#' It contains annual trade statistics starting from 1988
 #' and monthly trade statistics since 2000 for goods data
 #'
-#' Parameters that accept `NULL` will query all possible values. For example, setting `commodity_code = NULL`
-#' will retrieve data for all commodity codes. This can be useful for broad queries but may result in large datasets.
+#' Parameters that accept `NULL` will query all possible values.
+#' For example, setting `commodity_code = NULL`
+#' will retrieve data for all commodity codes.
+#' This can be useful for broad queries but may result in large datasets.
 #'
-#' @param frequency The frequency of returned trade data. Possible values: 'A' for annual data, 'M' for monthly data. Default: 'A'.
-#' @param type The type of returned trade data. Possible values: 'goods' for trade in goods, 'services' for trade in services. Default: 'goods'.
-#' @param commodity_classification The trade classification scheme. Possible values for goods: `c('HS','S1','S2','S3','S4','SS','B4','B5')`; for services: `c('EB02','EB10','EB10S','EB')`. Default: 'HS'.
-#' @param commodity_code The commodity code(s) or `NULL`. See `comtradr::ct_get_ref_table('HS')` for possible values. Default: 'TOTAL' (sum of all commodities).
-#' @param flow_direction The direction of trade flows or `NULL`. Possible values: 'import', 'export', 're-import', 're-export', 'all'. Default: 'all'.
-#' @param reporter Reporter ISO3 code(s) or `NULL`. See `comtradr::country_codes` for possible values. Default: 'all'.
-#' @param partner Partner ISO3 code(s) or `NULL`. See `comtradr::country_codes` for possible values. Default: 'World' (all partners as an aggregate).
-#' @param start_date The start date of the query. Format: `yyyy` for yearly, `yyyy-mm` for monthly.
-#' @param end_date The end date of the query. Format: `yyyy` for yearly, `yyyy-mm` for monthly. Max: 12 years after start date for annual data, one year for monthly data.
-#' @param primary_token Your primary UN Comtrade API token. Default: stored token from `comtradr::set_primary_comtrade_key`.
-#' @param process If TRUE, returns a data.frame with results. If FALSE, returns the raw httr2 request. Default: TRUE.
[…]
