+++ tests/testthat/test-ct_get_data.R
@@ -62,7 +62,7 @@ without_internet({
                   start_date = '2010',
                   end_date = '2010',
                   flow_direction = 'all',
-                  primary_token = 'test',
+                  # primary_token = 'test',
                   partner_2 = 'World',verbose = T)
               ,'https://comtradeapi.un.org/data/v1/get/S/A/EB?cmdCode=200&flowCode=M%2CX%2CRM%2CRX&partnerCode=32%2C280%2C276&reporterCode=156&period=2010&motCode=0&partner2Code=0&customsCode=C00&includeDesc=TRUE') # nolint
             })
@@ -99,7 +99,8 @@ httptest2::with_mock_dir("goods", {
                                           verbose = T,
                                           flow_direction = 'all',
                                           primary_token = 'test',
-                                          partner_2 = 'World'),'Got a response object from UN Comtrade. Use `process = F` if there is an error' ) # nolint
+                                          partner_2 = 'World'),
+    'Got a response object from UN Comtrade. Use `process = F` if there is an error' ) # nolint
   })
 })
 
