### 2023-12-15 eee4b80
deleted dependency vcr, which was unnecessary all along
