### 2023-12-15 2b4cf65
nord for replace_month function
