### 2023-12-15 b3dc1ce
add correct mock-dir with simplify=F so that headers are recorded
