+++ tests/testthat/test-ct_get_data.R
@@ -62,7 +62,7 @@ without_internet({
                   start_date = '2010',
                   end_date = '2010',
                   flow_direction = 'all',
-                  # primary_token = 'test',
+                  primary_token = 'test',
                   partner_2 = 'World',verbose = T)
               ,'https://comtradeapi.un.org/data/v1/get/S/A/EB?cmdCode=200&flowCode=M%2CX%2CRM%2CRX&partnerCode=32%2C280%2C276&reporterCode=156&period=2010&motCode=0&partner2Code=0&customsCode=C00&includeDesc=TRUE') # nolint
             })
@@ -87,7 +87,7 @@ without_internet({
             })
 })
 
-httptest2::with_mock_dir("goods", {
+httptest2::with_mock_dir("data", {
   test_that("test informative message about having gotten data", {
     expect_message(comtradr::ct_get_data(type = 'goods',
                                           commodity_classification = 'HS',
