+++ man/ct_build_request.Rd
@@ -9,9 +9,11 @@ ct_build_request(params, primary_token = NULL, verbose = FALSE)
 \arguments{
 \item{params}{a named vector of parameters for the comtrade request, result from \code{ct_check_params()}.}
 
-\item{primary_token}{Your primary UN Comtrade API token. Default: stored token from \code{comtradr::set_primary_comtrade_key}.}
+\item{primary_token}{Your primary UN Comtrade API token.
+Default: stored token from \code{comtradr::set_primary_comtrade_key}.}
 
-\item{verbose}{If TRUE, sends status updates to the console. If FALSE, runs functions quietly. Default: FALSE.}
+\item{verbose}{If TRUE, sends status updates to the console.
+If FALSE, runs functions quietly. Default: FALSE.}
 }
 \value{
 a httr2 request object
