+++ man/ct_get_data.Rd
@@ -18,71 +18,105 @@ ct_get_data(
   tidy_cols = TRUE,
   verbose = FALSE,
   primary_token = get_primary_comtrade_key(),
-  mode_of_transport = "0",
+  mode_of_transport = "TOTAL modes of transport",
   partner_2 = "World",
   customs_code = "C00",
   update = FALSE,
   requests_per_second = 10/60,
-  ...
+  extra_params = NULL
 )
 }
 \arguments{
-\item{type}{The type of returned trade data. Possible values: 'goods' for trade in goods, 'services' for trade in services. Default: 'goods'.}
+\item{type}{The type of returned trade data.
+Possible values: 'goods' for trade in goods,
+'services' for trade in services. Default: 'goods'.}
 
-\item{frequency}{The frequency of returned trade data. Possible values: 'A' for annual data, 'M' for monthly data. Default: 'A'.}
+\item{frequency}{The frequency of returned trade data.
+Possible values: 'A' for annual data, 'M' for monthly data. Default: 'A'.}
 
-\item{commodity_classification}{The trade classification scheme. Possible values for goods: \code{c('HS','S1','S2','S3','S4','SS','B4','B5')}; for services: \code{c('EB02','EB10','EB10S','EB')}. Default: 'HS'.}
+\item{commodity_classification}{The trade classification scheme.
+Possible values for goods: \code{c('HS','S1','S2','S3','S4','SS','B4','B5')};
+for services: \code{c('EB02','EB10','EB10S','EB')}. Default: 'HS'.}
 
-\item{commodity_code}{The commodity code(s) or \code{NULL}. See \code{comtradr::ct_get_ref_table('HS')} for possible values. Default: 'TOTAL' (sum of all commodities).}
+\item{commodity_code}{The commodity code(s) or \code{NULL}.
+See \code{comtradr::ct_get_ref_table('HS')} for possible values.
+Default: 'TOTAL' (sum of all commodities).}
 
-\item{flow_direction}{The direction of trade flows or \code{NULL}. Possible values: 'import', 'export', 're-import', 're-export', 'all'. Default: 'all'.}
+\item{flow_direction}{The direction of trade flows or \code{NULL}.
+Possible values: 'import', 'export', 're-import',
+'re-export', 'all'. Default: 'all'.}
 
-\item{reporter}{Reporter ISO3 code(s) or \code{NULL}. See \code{comtradr::country_codes} for possible values. Default: 'all'.}
+\item{reporter}{Reporter ISO3 code(s) or \code{NULL}.
+See \code{comtradr::country_codes} or \code{comtradr::ct_get_ref_table('reporter')}
+for possible values. Default: 'all'.}
 
-\item{partner}{Partner ISO3 code(s) or \code{NULL}. See \code{comtradr::country_codes} for possible values. Default: 'World' (all partners as an aggregate).}
+\item{partner}{Partner ISO3 code(s) or \code{NULL}. See \code{comtradr::country_codes}
+for possible values. Default: 'World' (all partners as an aggregate).}
 
-\item{start_date}{The start date of the query. Format: \code{yyyy} for yearly, \code{yyyy-mm} for monthly.}
+\item{start_date}{The start date of the query.
+Format: \code{yyyy} for yearly, \code{yyyy-mm} for monthly.}
 
-\item{end_date}{The end date of the query. Format: \code{yyyy} for yearly, \code{yyyy-mm} for monthly. Max: 12 years after start date for annual data, one year for monthly data.}
+\item{end_date}{The end date of the query.
+Format: \code{yyyy} for yearly, \code{yyyy-mm} for monthly.
+Max: 12 years after start date for annual data, one year for monthly data.}
 
[…]

[…]
