### 2023-12-15 9e714f4
deleted rest library call to vcr and unreferenced variable in check_mot
