+++ tests/testthat/setup.R
@@ -1,4 +1,3 @@
 library(httptest2)
 library(testthat)
 library(lifecycle)
-library(vcr)
