### 2023-12-17 b662fa4
implemented ct_get_ref_table for flow_direction, changed input parameters to text variable from comtrade and changed NULL parameter to everything. Default is still the same, so this should not confuse too much in the beginning.
