### 2023-12-17 c3af057
changed flow_direction parameter to case-insensitive to avoid confusion and make the update non-breaking for old code snippets (except for 'all')
