### 2023-12-19 #613 @potterzot
@datapumpernickel just a couple of thoughts on Q1 and Q2:

- Q1 I have run into this as a user of packages before as well. It can be tricky to debug because it's easy for your eyes to gloss over the mispelling of `partner`. In the `rnassqs` package, we handle it by checking all parameters for validity before querying. In this way a misspelled parameter will return an error that essentially says "Error: `patner` is not a valid parameter."

```
#' Expand a list handling null list as well
expand_list <- function(...){
  x <- list(...)
  if(length(x) == 0) { NULL } else { if(is.null(names(x))) x[[1]] else x }
}

#' Check validity of functions
parameter_is_valid <- function(param) {
  valid_params <- toupper(c(nassqs_params(), "param"))
  param2 <- gsub("__LE|__LT|__GT|__GE|__LIKE|__NOT_LIKE|__NE", "", toupper(param))

  if(!param2 %in% valid_params) {
    stop("Parameter '", param, "' is not a valid parameter. Use `nassqs_params()`
    for a list of valid parameters.")
  }
  return(invisible())
}

#' Make the query, checking parameter validity beforehand
nassqs_GET <- function(...,
                       api_path = c("api_GET",
                                    "get_param_values",
                                    "get_counts"),
                       progress_bar = TRUE,
                       format = c("csv", "json", "xml")) {

  # match args
  api_path <- match.arg(api_path)
  format <- match.arg(format)

  params <- expand_list(...)

  # Check that names of the parameters are in the valid parameter list
  for(x in names(params)) { parameter_is_valid(x) }

  ...
}
```

Here are links to the above functions to see the code in entirety: [expand_list()](https://github.com/ropensci/rnassqs/blob/26931d93ce84a9ddc624c5e9a8f175a9a07cd340/R/helpers.R#L13), [parameter_is_valid()](https://github.com/ropensci/rnassqs/blob/26931d93ce84a9ddc624c5e9a8f175a9a07cd340/R/helpers.R#L28), [nassqs_GET](https://github.com/ropensci/rnassqs/blob/26931d93ce84a9ddc624c5e9a8f175a9a07cd340/R/request.R#L224)

There is a [function](https://github.com/ropensci/rnassqs/blob/26931d93ce84a9ddc624c5e9a8f175a9a07cd340/R/params.R#L20) that helps with this that will just return all valid parameters and their description to help with usability as well.

- Q2: My inclination would be one of two options:
  - (1) set the default argument for all these parameters to `total` and to check in the function to return an error if the value is `NULL`. `total` would return the totals, `all` would return all subcategories but not the total, and `NULL` would return an error;
  - (2) set the default argument to `NULL`, which returns the total by default, and allow an `all` option that would return all values

In general, as a user I prefer a default behavior that fetches less data, so would not prefer the option of having a default that fetches all subcategory values.
