### 2023-12-21 c190e89
implement 'everything' as an alterantive to 'NULL" in documentation, functions, and tests
