+++ man/ct_perform_request.Rd
@@ -9,7 +9,7 @@ ct_perform_request(req, requests_per_second, verbose = FALSE)
 \arguments{
 \item{req}{a valid comtrade request built by the \code{ct_build_request()} function}
 
-\item{requests_per_second}{rate of requests per second executed,
+\item{requests_per_second}{Rate of requests per second executed,
 usually specified as a fraction, e.g. 10/60 for 10 requests per minute,
 see \code{req_throttle()} for details.}
 

[…]
