+++ vignettes/transition.Rmd
@@ -1,0 +1,201 @@
+---
+title: "Transition"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{transition}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+
+## Transitioning from the old API to the new API 🔄
+
+With the update of the Comtrade API by the United Nations, this package has
+undergone a comprehensive rewrite. Most functions that were available have been
+deprecated and there are breaking changes also in the names of arguments and 
+possible parameter values. 
+
+With the below examples, we hope to make the transition a little easier. Most 
+of the design principles of the package have remained similar. 
+
+The most important changes can be summarised in that the package: 
+- extensively checks parameters for validity before submitting 
+them
+- allows iso3 standardized country codes as inputs
+- queries the new parameters the UN added, such as a mode of transport.
+
+You will see that most other things have stayed more or less the same and the 
+transition will be a breeze! 💨
+
+### The basics 📊
+
+
+```{r, eval = FALSE}
+#### Previously
+q <- ct_search(reporters = "USA", 
+               partners = c("Germany", "France", "Japan", "Mexico"), 
+               trade_direction = "imports")
+
+#### Now
+q <- ct_get_data(reporter = "USA", 
+               partner = c("DEU", "FRA", "JPN", "MEX"), 
+               flow_direction = "import",
+               start_date = 2020,
+               end_date = 2023)
+```
+
+### The time parameter
+
+```{r, eval = FALSE}
+#### Previously
+# Get all monthly data for a single year (API max of 12 months per call).
+q <- ct_search(reporters = "USA", 
+               partners = c("Germany", "France", "Japan", "Mexico"), 
+               trade_direction = "imports", 
+               start_date = 2012, 
+               end_date = 2012, 
+               freq = "monthly")
+
+# Get monthly data for a specific span of months (API max of five months per call).
+q <- ct_search(reporters = "USA", 
+               partners = c("Germany", "France", "Japan", "Mexico"), 
+               trade_direction = "imports", 
+               start_date = "2012-03", 
+               end_date = "2012-07", 
+               freq = "monthly")
+
+
+#### Now
+# Get all monthly data for a single year (API max of 12 months per call).
+q <- ct_get_data(reporter = "USA", 
+               partner = c("DEU", "FRA", "JPN", "MEX"), 
+               flow_direction = "import",
+               start_date = 2012, 
+               end_date = 2012, 
+               frequency = "M"
+               )
+
+# Get monthly data for a specific span of months (API max of five months per call).
+q <- ct_get_data(reporter = "USA", 
+               partner = c("DEU", "FRA", "JPN", "MEX"), 
+               flow_direction = "import",
+               start_date = "2012-03", 
+               end_date = "2012-07", 
+               frequency = "M"
+               )
+```
+
+
+
+### Country Names 🌍
+
+#### Previously
+Countries passed to parameters `reporters` and `partners` must be spelled as they appear in the Comtrade country reference table. Function `ct_country_lookup` allows us to query the country reference table.
+```{r, eval = F}
+ct_country_lookup("korea", "reporter")
+ct_country_lookup("bolivia", "partner")
+q <- ct_search(reporters = "Rep. of Korea", 
+               partners = "Bolivia (Plurinational State of)", 
+               trade_direction = "all")
+```
+
+#### Now
+No need to specify the Comtrade country name, just use iso3 codes, as you can extract them from a myriad of other packages, e.g. `countrycodes`, `rnaturalearth` or `giscoR`.
+
+```{r, eval = F}
+
[…]
