### 2023-12-22 #613 @yabellini
Hi @datapumpernickel. rOpenSci Community Manager here. If you want to join our Slack, I can send an invitation for you.  Please let me know. You can write me to yabellini@ropensci.org
