### 2023-12-22 6d291ee
build vignette with example for large data
