+++ vignettes/large_data.Rmd
@@ -1,0 +1,374 @@
+---
+title: "Querying large amounts of data"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{large_data}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+```{r setup}
+library(comtradr)
+library(dplyr)
+library(ggplot2)
+
+```
+
+### The limits
+When you are on the free API tier, you will encounter two limits when wanting 
+to calculate on rather large datasets: 
+
+1) you can only make 250 calls per day, 
+
+2) of which each can take up to 100.000 rows. 
+
+So when trying to fetch the maximum amount of data in the minimum amount of 
+days, you want to minimize the amount of requests you need by making them as comprehensive as possible, without ever exceeding the 100k row limit. 
+
+### The example - Imports of the EU
+
+Let's say, we want to know which is the top exporter of the EU in each product 
+class of the Harmonic System. This could be useful, when thinking about 
+dependencies between countries, e.g. when evaluating the usefulness or impact of Regulation such as the 
+[EUDR Regulation](https://environment.ec.europa.eu/topics/forests/deforestation/regulation-deforestation-free-products_en). Quoting from the [ITC](https://tradebriefs.intracen.org/2023/11/spotlight): 
+"Under the regulation that entered into force on 29 June 2023, any 
+operator or trader placing [specific] commodities on the EU market or 
+exporting them from the EU, has to be able to prove that the goods do not 
+originate from deforested land (cutoff date 31 December 2020) or contribute 
+to forest degradation". One question that arises would be, what are the 
+impacts of this regulation on other countries and which of these is 
+particularly important to the EU?
+
+Let's replicate some of their numbers, e.g. the relevance to the EU 
+indicator from the map of the ITC spotlight cited above. 
+
+*What is the share of one country in the EU's imports of a product?*
+
+Remember, to get Comtrade data, we need the Commodity Codes, 
+the respective iso3 codes for the countries and a time frame. 
+
+#### 1. What are the commodity codes?
+
+First we need to find out, which HS codes, or which goods are affected. 
+
+We look at Annex 1 of the [Regulation](https://eur-lex.europa.eu/legal-content/EN/TXT/?uri=CELEX%3A32023R1115&qid=1687867231461). 
+
+This gives us more or less the following list of commodities for the 
+product class 'Wood'. 
+
+```{r}
+wood <-
+  c(
+    "4401",
+    "4402",
+    "4403",
+    "4404",
+    "4405",
+    "4406",
+    "4407",
+    "4408",
+    "4409",
+    "4410",
+    "4411",
+    "4412",
+    "4413",
+    "4414",
+    "4415",
+    "4416",
+    "4417",
+    "4418",
+    "4419",
+    "4420",
+    "4421",
+    "940330",
+    "940340",
+    "940350",
+    "940360",
+    "940391",
+    "940610",
+    "4701",
+    "4702",
+    "4703",
+    "4704",
+    "4705",
+    "470610", 
+    "470691", 
+    "470692", 
+    "470693",
+    "48",
+    "49",
+    "9401"
+  )
+wood_df <- data.frame(cmd_code = wood, product_cat = "wood")
+
+```
+
+#### 2. Which are the countries?
+
+First we need a list of the EU27 countries as iso3 codes. 
+I use the `giscoR` package to get this information quickly. 
+
+```{r, echo = FALSE}
+v_data_6 <- system.file("extdata", "vignette_data_6.rda", package = "comtradr")
[…]
