### 2023-12-22 0e059c7
get data for vignettes
