### 2023-12-22 e70be60
updated vignettes with some spelluing and transition info
