+++ README.Rmd
@@ -24,7 +24,6 @@ Interface with and extract data from the United Nations Comtrade API <https://co
 
 The Comtrade API has been undergoing extensive updates. At this point the legacy API has been taken offline (see [here](https://unstats.un.org/wiki/display/comtrade/New+Comtrade+FAQ+for+Advanced+Users#NewComtradeFAQforAdvancedUsers-WhatisthelegacyoftheUNComtrade?UntilwhencanIuseit?)).
 
-
 In order to accommodate the new syntax, as well as adding the new functionality, the `comtradr` package has been undergoing an extensive re-write. Additionally, it is not possible anymore to query the API without an API token. There still is a free version, but unlike before you need to be a registered user to obtain the token. See the [FAQ](https://unstats.un.org/wiki/display/comtrade/New+Comtrade+User+Guide#NewComtradeUserGuide-UNComtradeAPIManagement) for details on how to obtain the new access tokens. 
 
 Please [report](https://github.com/ropensci/comtradr/issues) issues, comments, or feature requests. We are very much looking for feedback on the usability of the new functions. 
@@ -54,7 +53,21 @@ install.packages("comtradr")
 
 ### Authentication 🔐
 
-As stated above, you need an API token, see the [FAQ](https://unstats.un.org/wiki/display/comtrade/New+Comtrade+User+Guide#NewComtradeUserGuide-UNComtradeAPIManagement) for details on how to obtain the access tokens. 
+**Do not be discouraged by the complicated access to the token 
+- you can do it! 💪**
+
+As stated above, you need an API token, see the FAQ of Comtrade for details
+on how to obtain it:
+
+➡️ https://unstats.un.org/wiki/display/comtrade/New+Comtrade+User+Guide#NewComtradeUserGuide-UNComtradeAPIManagement 
+
+You need to follow the detailed explanations,
+which include screenshots, in the Wiki of Comtrade to the letter. ☝️
+I am not writing them out here, because they might be updated regularly. 
+However, once you are signed up, select the `comtrade - v1` 
+product, which is the free API. 
+
+#### Storing the API key
 
 If you are in an interactive session, you can call the following function to save your API token to the environment file for the current session. 
 
@@ -64,15 +77,19 @@ library(comtradr)
 set_primary_comtrade_key()
 ```
 
-If you are not in an interactive session, you can register the token once in your session using the following base-r function.
+If you are not in an interactive session, you can register the token once in 
+your session using the following base-r function.
 
 ```{r, eval = F}
 Sys.setenv('COMTRADE_PRIMARY' = 'xxxxxxxxxxxxxxxxx')
 ```
 
-If you would like to set the comtrade key permanently, we recommend editing the project `.Renviron` file, where you need to add a line with `COMTRADE_PRIMARY = xxxx-your-key-xxxx`. 
+If you would like to set the comtrade key permanently, we recommend editing 
+the project `.Renviron` file, where you need to add a line 
+with `COMTRADE_PRIMARY = xxxx-your-key-xxxx`. 
 
-ℹ️ Do not forget the line break after the last entry. This is the easiest by taking advantage of the great `usethis` package. 
+ℹ️ Do not forget the line break after the last entry. This is the 
+easiest by taking advantage of the great `usethis` package. 
 
 ```{r, eval = F}
 usethis::edit_r_environ(scope = 'project')
@@ -80,7 +97,8 @@ usethis::edit_r_environ(scope = 'project')
 
 ### Example 1 ⛴️
 
-Now we can get to actually request some data. Let us query the total trade between China and Germany and Argentina, as reported by China. 
+Now we can get to actually request some data. Let us query the total 
[…]
