+++ R/utils.R
@@ -8,7 +8,7 @@
 #' @export
 #' @examplesIf interactive()
 #' ## set API key
-#' set_primary_comtrade_key('xxxxxc678ca4dbxxxxxxxx8285r3')
+#' set_primary_comtrade_key("xxxxxc678ca4dbxxxxxxxx8285r3")
 #'
 set_primary_comtrade_key <- function(key = NULL) {
   if (is.null(key)) {
@@ -83,26 +83,25 @@ get_primary_comtrade_key <- function() {
 #' ct_get_ref_table("reporter")
 #'
 ct_get_ref_table <- function(dataset_id, update = FALSE, verbose = FALSE) {
-
   ## make switch to the name of the datasets, which are slightly different to the dataset_ids
   switch_list <- c(
-    'B4'    = 'cmd_b4'   ,
-    'B5'    = 'cmd_b5'   ,
-    'EB02'  = 'cmd_eb02' ,
-    'EB10'  = 'cmd_eb10' ,
-    'EB10S' = 'cmd_eb10s',
-    'EB'    = 'cmd_eb'   ,
-    'HS'    = 'cmd_hs'   ,
-    'S1'    = 'cmd_s1'   ,
-    'S2'    = 'cmd_s2'   ,
-    'S3'    = 'cmd_s3'   ,
-    'S4'    = 'cmd_s4'   ,
-    'SS'    = 'cmd_ss'   ,
-    'reporter'    = 'reporter'   ,
-    'partner'    = 'partner',
-    'mode_of_transport'    = 'mot',
-    'flow_direction'    = 'flow',
-    'customs_code'    = 'customs'
+    "B4" = "cmd_b4",
+    "B5" = "cmd_b5",
+    "EB02" = "cmd_eb02",
+    "EB10" = "cmd_eb10",
+    "EB10S" = "cmd_eb10s",
+    "EB" = "cmd_eb",
+    "HS" = "cmd_hs",
+    "S1" = "cmd_s1",
+    "S2" = "cmd_s2",
+    "S3" = "cmd_s3",
+    "S4" = "cmd_s4",
+    "SS" = "cmd_ss",
+    "reporter" = "reporter",
+    "partner" = "partner",
+    "mode_of_transport" = "mot",
+    "flow_direction" = "flow",
+    "customs_code" = "customs"
   )
 
   ## check dataset id for valid values
@@ -116,64 +115,73 @@ ct_get_ref_table <- function(dataset_id, update = FALSE, verbose = FALSE) {
 
   ## if the dataset is not yet loaded into the environment
   ## read it from disk and save to environment
-  if(is.null(data)){
-    data <- fs::path_package(paste0('extdata/',ref_table_id,'.rds'),
-                             package = 'comtradr') |>
+  if (is.null(data)) {
+    data <- fs::path_package(paste0("extdata/", ref_table_id, ".rds"),
+      package = "comtradr"
+    ) |>
       readr::read_rds()
-    assign(dataset_id,data,envir = ct_env)
+    assign(dataset_id, data, envir = ct_env)
   }
 
-  if(update & any(dataset_id %in% ct_env$updated)){
+  if (update & any(dataset_id %in% ct_env$updated)) {
     ## if update is true, but dataset_id has already been updated once
     ## only return message
     if (verbose) {
-      cli::cli_inform(c("i" = paste0("Already checked for updates for ",
-                                     dataset_id,' in this session.')))
+      cli::cli_inform(c("i" = paste0(
+        "Already checked for updates for ",
+        dataset_id, " in this session."
+      )))
     }
     return(data)
-  } else if(update){
+  } else if (update) {
     ## if update is true and not yet updated in this session inform user that update process is starting
     if (verbose) {
[…]
