+++ R/ct_process_response.R
@@ -1,8 +1,12 @@
 #' Processes the response object
 #'
-#' The function is internally called by `ct_get_data()` and processes the httr2response object returned by `ct_perform_request()`, it parses the JSON and adds the respective ISO codes for the reporter and partner countries, as well as the commodity code description.
+#' The function is internally called by `ct_get_data()` and processes
+#' the httr2response object returned by `ct_perform_request()`, it parses the
+#' JSON and adds the respective ISO codes for the reporter and partner
+#' countries, as well as the commodity code description.
 #'
-#' @param resp a valid httr2 response object created from the function `ct_perform_request()`
+#' @param resp a valid httr2 response object created
+#' from the function `ct_perform_request()`
 #'
 #' @returns a data.frame object with the results
 #'
@@ -16,7 +20,8 @@
 #'   flow_direction = "import"
 #' )
 #' req <- httr2::last_request()
-#' resp <- comtradr:::ct_perform_request(req, requests_per_second = 10 / 60, verbose = FALSE)
+#' resp <- comtradr:::ct_perform_request(req, requests_per_second = 10 / 60,
+#' verbose = FALSE)
 #' data <- comtradr:::ct_process_response(resp)
 #' @inheritParams ct_get_data
 ct_process_response <- function(resp, verbose = FALSE, tidy_cols) {
