### 2023-12-22 94f0da2
more tests for check_params, fix for motcode where something qwas duplicated
