### 2023-12-23 2dcd21c
fixed some minor errors in documentation of vignettes
