+++ man/get_primary_comtrade_key.Rd
@@ -7,10 +7,13 @@
 get_primary_comtrade_key()
 }
 \value{
-Gets your primary comtrade key from the environment var COMTRADE_PRIMARY
+Gets your primary comtrade key from the environment
+var COMTRADE_PRIMARY
 }
 \description{
-If you would like your Comtrade API key to persist in between sessions, use \code{usethis::edit_r_environ()} to add the env variable COMTRADE_PRIMARY to your environment file.
+If you would like your Comtrade API key to persist in between sessions,
+use \code{usethis::edit_r_environ()} to add the env variable COMTRADE_PRIMARY
+to your environment file.
 }
 \examples{
 \dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
