+++ man/ct_commodity_db_type.Rd
@@ -7,13 +7,15 @@
 ct_commodity_db_type(...)
 }
 \arguments{
-\item{...}{Used to catch all possible arguments that users have supplied to this function.}
+\item{...}{Used to catch all possible arguments that users
+have supplied to this function.}
 }
 \value{
 depreciation error
 }
 \description{
-This function is deprecated. There is currently no alternative for this function.
+This function is deprecated. There is currently
+no alternative for this function.
 \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
 }
 \examples{
