+++ man/ct_use_pretty_cols.Rd
@@ -7,13 +7,15 @@
 ct_use_pretty_cols(...)
 }
 \arguments{
-\item{...}{Used to catch all possible arguments that users have supplied to this function.}
+\item{...}{Used to catch all possible arguments that users
+have supplied to this function.}
 }
 \value{
 depreciation error
 }
 \description{
-This function is deprecated. Please use the \code{process} argument in the main function instead.
+This function is deprecated. Please use the \code{process}
+argument in the main function instead.
 \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
 }
 \examples{
