+++ man/ct_process_response.Rd
@@ -7,7 +7,8 @@
 ct_process_response(resp, verbose = FALSE, tidy_cols)
 }
 \arguments{
-\item{resp}{a valid httr2 response object created from the function \code{ct_perform_request()}}
+\item{resp}{a valid httr2 response object created
+from the function \code{ct_perform_request()}}
 
 \item{verbose}{If TRUE, sends status updates to the console.
 If FALSE, runs functions quietly. Default: FALSE.}
@@ -19,7 +20,10 @@ If FALSE, returns raw column names. Default: TRUE.}
 a data.frame object with the results
 }
 \description{
-The function is internally called by \code{ct_get_data()} and processes the httr2response object returned by \code{ct_perform_request()}, it parses the JSON and adds the respective ISO codes for the reporter and partner countries, as well as the commodity code description.
+The function is internally called by \code{ct_get_data()} and processes
+the httr2response object returned by \code{ct_perform_request()}, it parses the
+JSON and adds the respective ISO codes for the reporter and partner
+countries, as well as the commodity code description.
 }
 \examples{
 \dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
@@ -32,7 +36,8 @@ ct_get_data(
   flow_direction = "import"
 )
 req <- httr2::last_request()
-resp <- comtradr:::ct_perform_request(req, requests_per_second = 10 / 60, verbose = FALSE)
+resp <- comtradr:::ct_perform_request(req, requests_per_second = 10 / 60,
+verbose = FALSE)
 data <- comtradr:::ct_process_response(resp)
 \dontshow{\}) # examplesIf}
 }
