+++ man/ct_update_databases.Rd
@@ -7,13 +7,15 @@
 ct_update_databases(...)
 }
 \arguments{
-\item{...}{Used to catch all possible arguments that users have supplied to this function.}
+\item{...}{Used to catch all possible arguments that users
+have supplied to this function.}
 }
 \value{
 depreciation error
 }
 \description{
-This function is deprecated. Please use  \code{update} parameter in the main \code{ct_get_data} function instead.
+This function is deprecated. Please use  \code{update}
+parameter in the main \code{ct_get_data} function instead.
 \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
 }
 \examples{
