### 2023-12-23 #613 @datapumpernickel
Dear @ernestguevarra, @potterzot and @noamross,

I have now implemented all changes to the best of my abilities. Thank you also for the helpful feedback on my additional questions. This has introduced a minor breaking change, as now `all` is not available anymore. But since the package is not on CRAN yet and the change is beneficial, I think that is ok.

I have basically kept the Defaults as restrictive as is, to not return too much data. There is no more NULL option, instead you can pass `everything` and in the case of `reporter` and `partner` you can also pass `all_countries` (formerly `all`), which makes it clear that not every parameter, but only countries are returned.

I have also gotten rid of `...` in favor of `extra_params` which now makes the package more robust for misspelled arguments - thanks also for your input potterzot!

Implemented changes are listed here: https://github.com/ropensci/comtradr/issues/69

As a next step, I would publish to CRAN, I suppose.

All the best and thank you again to everyone for the helpful comments and your time!
