### 2024-01-09 #613 @ernestguevarra
@datapumpernickel sorry for dropping the ball on this.

Thank you for all the progress achieved since my review. I agree with the changes made in relation to my review and also those that have been made in response to @potterzot and @noamross feedback.

I've worked with the latest version a couple of days ago and specifically tried the functions to which changes have been made. All looks good and I agree that the package is ready!

Really appreciate your efforts into this. I think a lot of people will benefit from this!
