### 2024-01-09 #613 @ropensci-review-bot
Approved! Thanks @datapumpernickel for submitting and @ernestguevarra, @potterzot for your reviews! :grin:
