### 2024-01-09 #613 @noamross
Thank you @datapumpernickel for bringing this package to this point and putting it through review after taking over maintenance!  Of course most of the steps above don't apply for a package already in our suite.

I do think that the story of this package and review would make a great blog post.  Taking over maintenance is a big thing to tackle and the [more stories](https://blogguide.ropensci.org/authorcontent.html) we have of it the better we are able to guide people through it! Please let us know if you are interested in writing one.
