### 2024-01-10 #613 @datapumpernickel
Hi all!

Really excited to see this package fully approved back in the fold of rOpenSci! Thanks to everybody for the input and your time.
I am in general interested in writing a blog article, but I am not sure when I fill find the time to do so.

Currently I am waiting for some feedback on the caching ability I am implementing. Then I would prioritize getting the package on CRAN. As soon as that is done I will be in touch about writing up the process! :)
